refactor: navigation into folder
